requirements.txt
PyQt6>=6.4.0
websockets>=10.4
python-dotenv>=1.0.0
mysql-connector-python>=8.0.32
bcrypt>=4.0.1
darkdetect>=0.8.0
cryptography>=40.0.0
